jaundice-backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
tensorflow==2.14.0
keras==2.14.0
numpy==1.24.3
pillow==10.1.0
opencv-python==4.8.1.78
scikit-learn==1.3.2
scipy==1.11.4
matplotlib==3.8.1
